requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py
pillow
asyncio
yarl==1.1.1
aiohttp<3.0.0
